* Reverted moving Options to Core and back to Api again. * Holiday is now HolidayService in Api
